+++ src/main/kotlin/dev/slint/ideaplugin/ide/settings/SlintLspSettings.kt
@@ -10,6 +10,7 @@ data class SlintLspSettings(
     var noToolbar: Boolean = false,
     var includePaths: MutableList<String> = mutableListOf(),
     var useExternalLsp: Boolean = false,
+    var providedByEditor: Boolean = false
 )
 
 enum class SlintStyle {